requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.4.1
asttokens==3.0.0
attrs==24.2.0
autopep8==2.3.1
av==12.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
dlib==19.24.6
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema==2.21.1
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
idna==3.10
ifaddr==0.2.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
narwhals==1.9.4
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
numpy==2.1.2
odfpy==1.4.1
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.52
pickleshare==0.7.5
pika==1.3.2
pillow==10.4.0
pip-check-reqs==2.5.3
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.28.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycodestyle==2.12.1
pycparser==2.22
pydeck==0.9.1
pyee==12.1.1
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.3.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.41.1
streamlit-keyup==0.2.4
streamlit-option-menu==0.4.0
streamlit-webrtc==0.47.9
tenacity==9.0.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validate-docbr==1.10.0
vulture==2.14
watchdog==5.0.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
wheel==0.44.0
yarg==0.1.9
